tutoring_website/requirements.txt
asgiref==3.5.2
crispy-bootstrap5==0.6
Django==4.1
django-crispy-forms==1.14.0
gunicorn==20.1.0
psycopg2==2.9.5
python-dotenv==0.21.0
sqlparse==0.4.2
